+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## 1.7.0.22 [Work-In-Progress]
 - Simplified source code:
   * expressions, slider names, etc. are now loaded directly in the PARM resource).
-	* Load-methods don't return boolean and a reason string. Instead they return 0 (success) or a error message ID otherwise.
-- As side effect: The slider positions will be preserved when a filter is made
+	* Load-methods and Save-methods don't return boolean and a reason string. Instead they return 0 (success) or a error message ID otherwise.
+- The slider positions will be preserved when a filter is made
 
 ## 1.7.0.21 [23-Nov-2023]
 - Read FFX file: Fixed buffer overflow when some strings (Title,Category,Author,Copyright,SliderNames) are too long